# Reduce web_server loop overhead on ESP32 by avoiding unnecessary semaphore operations