# What does this implement/fix?

This PR optimizes the web_server component loop on ESP32 by reducing unnecessary semaphore operations. Previously, the loop would take a semaphore lock on every iteration to check if there was work to do, even when the queue was empty most of the time. This caused measurable overhead in runtime statistics.

The fix adds an atomic flag to track whether the deferred work queue has items, allowing the loop to skip the expensive semaphore operation when the queue is empty. 

**Performance improvement measured:**
- Before: web_server used 620ms of CPU time over 60 seconds (avg 0.09ms per call)
- After: web_server uses only 1ms of CPU time over 60 seconds (avg 0.00ms per call)
- **Result: 99.8% reduction in CPU time spent in web_server loop**

This optimization eliminates the constant semaphore checking overhead while maintaining full functionality when there's actual work to process.

## Types of changes

- [x] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] Code quality improvements to existing code or addition of tests
- [ ] Other

**Related issue or feature (if applicable):**

N/A

**Pull request in [esphome-docs](https://github.com/esphome/esphome-docs) with documentation (if applicable):**

N/A - No documentation changes needed

## Test Environment

- [x] ESP32
- [x] ESP32 IDF
- [ ] ESP8266
- [ ] RP2040
- [ ] BK72xx
- [ ] RTL87xx

## Example entry for `config.yaml`:

```yaml
# No configuration changes required
web_server:
  port: 80
```

## Checklist:
  - [x] The code change is tested and works locally.
  - [ ] Tests have been added to verify that the new code works (under `tests/` folder).

If user exposed functionality or configuration variables are added/changed:
  - [ ] Documentation added/updated in [esphome-docs](https://github.com/esphome/esphome-docs).